Move canonicalization to separate file - document and split into functions
